Move EventSource to shared CoreLib partition
